+++ ready/AzNamingTool/Models/ResourceNameRequest.cs
@@ -1,17 +1,18 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace AzureNamingTool.Models
+﻿namespace AzureNamingTool.Models
 {
-    public class ResourceNameRequestWithComponents
+    public class ResourceNameRequest
     {
-        public ResourceDelimiter ResourceDelimiter { get; set; }
-        public ResourceEnvironment? ResourceEnvironment { get; set; }
-        public ResourceFunction? ResourceFunction { get; set; }
-        public string ResourceInstance { get; set; }
-        public ResourceLocation? ResourceLocation { get; set; }
-        public ResourceOrg? ResourceOrg { get; set; }
-        public ResourceProjAppSvc? ResourceProjAppSvc { get; set; }
-        public ResourceType ResourceType { get; set; }
-        public ResourceUnitDept? ResourceUnitDept { get; set; }
+        public string? ResourceEnvironment { get; set; }
+        public string? ResourceFunction { get; set; }
+        public string? ResourceInstance { get; set; }
+        public string? ResourceLocation { get; set; }
+        public string? ResourceOrg { get; set; }
+        public string? ResourceProjAppSvc { get; set; }
+        public string ResourceType { get; set; }
+        public string? ResourceUnitDept { get; set; }
+        /// <summary>
+        /// Dictionary [Custom Component Type Name],[Custom Component Short Name Value]
+        /// </summary>
+        public Dictionary<string, string>? CustomComponents { get; set; }
     }
 }
